make it better

+++ Controller/blogController.js
@@ -11,20 +11,28 @@ import {
 import { validateBlog } from "../Common/validateBlog.js";
 import uploadFile from "../Common/upload.js";
 import deleteBlogImage from "../Common/deleteBlogImage.js";
+import { removeBlogComments } from "../Services/blogCommentsService.js";
 
-export const getBlogs = async (req, res) => {
-  // console.log("Getting Blogs")
+// Get Blogs For Home Page
+export const getHomeBlogs = async (req, res) => {
   try {
     const blogs = await findHomeBlogs();
-    if (blogs.success) {
-      return res.status(200).json({
-        data: blogs,
-        success: true,
-        error: false,
-      });
-    } else {
-      throw new Error("Failed to Load Blogs");
+
+    // Check if Error Occurred while fetching Blogs
+    if (blogs?.serverError) {
+      return res.status(500).json({
+        success: false,
+        serverError: true
+      })
     }
+    res.status(200).json({
+      error: false,
+      success: true,
+      blogs: {
+        mainBlog: blogs.mainBlog,
+        homeBlogs: blogs.homeBlogsData
+      }
+    })
   } catch (error) {
     console.log(error.message);
     res.status(500).json({
@@ -34,10 +42,16 @@ export const getBlogs = async (req, res) => {
     });
   }
 };
+
+// Get Blogs Based on User
 export const getUserBlogs = async (req, res) => {
   const userId = req.params.userId;
   try {
+
+    // Validate UserId
     const validUser = await findUser(userId);
+
+    // If userId is invalid
     if (!validUser) {
       return res.status(400).json({
         error: true,
@@ -45,13 +59,27 @@ export const getUserBlogs = async (req, res) => {
         message: "User Does Not Exists",
       });
     }
+
+    // If error occurred while Validating User Id
+    if (validUser?.serverError) {
+      return res.status(500).json({
+        serverError: true,
+        success: false,
+      });
+    }
+
+    // Find Blogs based on User
     const findUserBlogs = await findBlogsByUser(userId);
+
+    // If Error Occured While Fetching Blogs
     if (!findUserBlogs.success) {
       return res.status(500).json({
         serverError: true,
         success: false,
       });
     }
+
+    // Success
     res.status(200).json({
       error: false,
       success: true,
@@ -65,15 +93,17 @@ export const getUserBlogs = async (req, res) => {
     });
   }
 };
+
+// Get a Particular blog based on Id
 export const getBlog = async (req, res) => {
-  // console.log("Getting Single Blog")
   try {
-    let sameUser = false;
+    let sameUser = false;  // Same user is True if the Blog Requested is Created by the User making this request
     const { id } = req.params;
     const user = req.user;
 
-    const { strict } = req.body
-    // if Strict set to true than User Must Be Logged in
+    const { strict } = req.body // if Strict set to true than User Must Be Logged in
+
+    // If Strict is true and User is not present
     if (strict && !user) {
       return res.status(401).json({
         error: true,
@@ -83,8 +113,10 @@ export const getBlog = async (req, res) => {
       })
     }
 
+    // Find Blog Based on Id
     const blogData = await findBlog(id);
-    // console.log(blogData);
+
+    // If Blog Does not Exists
     if (!blogData.blog) {
       return res.status(404).json({
         error: true,
@@ -94,25 +126,45 @@ export const getBlog = async (req, res) => {
         blog: false,
       });
     }
+
+    // Error occurred while finding Blog
+    if (blogData?.serverError) {
+      return res.status(500).json({
+        serverError: true,
+        success: false,
+
+      });
+    }
+
+    // If userId (present in the parametre) is same as the Blog Creator Id
     if (blogData.blog.creator._id == user) {
       const checkUser = await findUser(user);
       if (checkUser) {
         sameUser = true;
       }
     }
+
+    // Check is Blog is liked by the user
+    let liked = false
+    if (user) {
+      if (blogData.blog.likes.includes(user)) {
+        liked = true
+      }
+    }
+
     if (!sameUser) {
       return res.status(200).json({
         error: false,
         success: true,
-        blog: blogData.blog,
-        sameUser: false,
+        blog: { ...blogData.blog.toObject(), liked },
+        sameUser: false
       });
     } else {
       return res.status(200).json({
         error: false,
         success: true,
-        blog: blogData.blog,
-        sameUser: true,
+        blog: { ...blogData.blog.toObject(), liked },
+        sameUser: true
       });
     }
   } catch (error) {
@@ -125,18 +177,19 @@ export const getBlog = async (req, res) => {
   }
 };
 
+// Create a New Blog
 export const createNewBlog = async (req, res) => {
   try {
     //  User from JWT Token
     const userId = req.user;
     // Form Data
-    const { title, description, content, tags, category, otherCategory } =
+    const { title, description, content, tags, category, subCategory, otherCategory } =
       req.body;
+
     // File from Front-end
     const file = req.file;
-    console.log("file is", file)
 
-
+    // If User Not Present
     if (!userId) {
       return res.status(401).json({
         error: true,
@@ -148,8 +201,11 @@ export const createNewBlog = async (req, res) => {
         }
       });
     }
+
     // Validate User
     const isUser = await findUser(userId);
+
+    // If UserId provided is invalid
     if (!isUser) {
       return res.status(401).json({
         error: true,
@@ -162,7 +218,15 @@ export const createNewBlog = async (req, res) => {
       });
     }
 
-    // Validate Blog
+    // If error occurred while Validating User Id
+    if (isUser?.serverError) {
+      return res.status(500).json({
+        serverError: true,
+        success: false,
+      });
+    }
+
+    // Validate Blog - Blog Validation is first done Manually, because Blog Image then must be uploaded to 3rd Party Cloud Storage
     const validate = validateBlog({
       title,
       content,
@@ -170,11 +234,12 @@ export const createNewBlog = async (req, res) => {
       tags,
       category,
       otherCategory,
+      subCategory,
       file,
       noFile: false,
     });
-    console.log(validate);
-    // Send Error if Validation Failed
+
+    // Send Error if Validation Failed 
     if (validate.error) {
       return res.status(400).json({
         success: false,
@@ -182,34 +247,41 @@ export const createNewBlog = async (req, res) => {
         formError: validate.form,
       });
     }
-    console.log("File PAth is ", file.path)
+
+    // Upload Blog Image to 3rd Part Cloud Storage (Cloudinary)
     const fileUpload = await uploadFile(file.path, process.env.BLOG_MAIN_IMAGE_UPLOAD_PRESET_CLOUDINARY);
-    console.log(fileUpload);
     if (fileUpload.serverError) {
       return res.status(500).json({
         success: false,
         serverError: true,
       });
     }
+
+    // Save the Blog and the Image Url to the Database
     const newBlog = await createBlog({
       title,
       description,
       content,
       category,
       otherCategory,
+      subCategory,
       tags: tags.split(" "),
       imageUrl: fileUpload.url,
       likes: [],
       comments: [],
       creator: userId,
+      likesCount: 0
     });
-    // console.log(newBlog);
+
+    // if Error Occurred while Saving data to the database
     if (newBlog.serverError) {
       return res.status(500).json({
         success: false,
         serverError: true,
       });
     }
+
+    // Check for Validation Errors
     if (newBlog.formError) {
       return res.status(400).json({
         error: true,
@@ -230,10 +302,11 @@ export const createNewBlog = async (req, res) => {
   }
 };
 
+// Delete a Blog
 export const deleteBlog = async (req, res) => {
   try {
-    const user = req.user;
-    const { id } = req.params;
+    const user = req.user;  // User Id that is sent from the Auth Middleware
+    const { id } = req.params; // BlogId
     if (!user) {
       return res.status(401).json({
         error: true,
@@ -242,7 +315,11 @@ export const deleteBlog = async (req, res) => {
         auth: false,
       });
     }
+
+    // Find Blog based on the id provided
     const validBlog = await findBlog(id);
+
+    // if Blog does not exist
     if (!validBlog.blog) {
       return res.status(400).json({
         error: true,
@@ -251,6 +328,16 @@ export const deleteBlog = async (req, res) => {
         blog: false,
       });
     }
+
+    // if error occurred while Finding Blog
+    if (validBlog?.serverError) {
+      return res.status(500).json({
+        serverError: true,
+        success: false,
+      });
+    }
+
+    // If Blog Creator id is different from the userId (Requesting User)
     if (validBlog.blog.creator._id != user) {
       return res.status(401).json({
         error: true,
@@ -259,10 +346,28 @@ export const deleteBlog = async (req, res) => {
         auth: false,
       });
     }
+
+    // Image URl
     let fileUrl = validBlog.blog.imageUrl
-    // console.log(fileUrl)
-    await removeBlog(id);
+
+    // Remove Blog
+    const permanentDelete = await removeBlog(id);
+
+    // If error Occurred While Deleting Blog
+    if(permanentDelete?.serverError){
+      return res.status(500).json({
+        serverError:true,
+        success:false
+      })
+    }
+
+    // Delete Blog Comments
+    await removeBlogComments(id)
+
+    //  Delete Blog Image from 3rd party Cloud Storage (Cloudinary)
     deleteBlogImage(fileUrl)
+
+    // Success
     res.status(200).json({
       message: "Blog Deleted Successfully",
       success: true,
@@ -271,19 +376,21 @@ export const deleteBlog = async (req, res) => {
     console.log(error.message);
     res.status(500).json({
       serverError: true,
-      message: "Something Went Wrong, Please Try Again.",
       success: false,
     });
   }
 };
 
+// Update a Blog
 export const updateBlog = async (req, res) => {
   try {
     const { blogId } = req.params;
     const user = req.user;
     const file = req.file;
-    const { title, description, content, tags, category, otherCategory } =
+    const { title, description, content, tags, category, subCategory, otherCategory } =
       req.body;
+
+    // If User is not Present
     if (!user) {
       return res.status(401).json({
         error: true,
@@ -296,7 +403,11 @@ export const updateBlog = async (req, res) => {
         }
       });
     }
+
+    // Find Blog based on the BlogID
     const validBlog = await findBlog(blogId);
+
+    // If Blog Not Found
     if (!validBlog.blog) {
       return res.status(400).json({
         error: true,
@@ -308,6 +419,16 @@ export const updateBlog = async (req, res) => {
         }
       });
     }
+
+    // If Error Occurred while Finding Blog
+    if (validBlog.serverError) {
+      return res.status(500).json({
+        serverError: true,
+        success: false,
+      });
+    }
+
+    // If BlogCreatorId is not equals to the userId
     if (validBlog.blog.creator._id != user) {
       return res.status(401).json({
         error: true,
@@ -319,19 +440,21 @@ export const updateBlog = async (req, res) => {
         }
       });
     }
-    // Validate Blog
-
-    // Validate Blog
+   
+    // Validate Blog - Blog Validation s done Manually First to upload the updated Blog image (if present) to third party cloud storage
     const validate = validateBlog({
       title,
       content,
       description,
       tags,
       category,
+      subCategory,
       otherCategory,
       file,
       noFile: file ? false : true,
     });
+
+    // if Validdation Error
     if (validate.error) {
       return res.status(400).json({
         success: false,
@@ -339,10 +462,12 @@ export const updateBlog = async (req, res) => {
         formError: validate.form,
       });
     }
-    // console.log(validate)
+
+    // If File Present Upload it's url to Cloudinary (3rd Party Cloud Storage)
     let fileUrl;
     if (file) {
       const fileUpload = await uploadFile(file.path, process.env.BLOG_MAIN_IMAGE_UPLOAD_PRESET_CLOUDINARY);
+      // If Error While Uploading File 
       if (fileUpload.serverError) {
         return res.status(500).json({
           success: false,
@@ -351,28 +476,40 @@ export const updateBlog = async (req, res) => {
       }
       fileUrl = fileUpload.url;
     }
+
     // Delete the previous Blog Image if the User wants to Update it with a newer Image
     if (file) {
       const previousFileUrl = validBlog.blog.imageUrl
       deleteBlogImage(previousFileUrl)
     }
+
+    // Update Blog Data
     const BlogDetails = {
       title,
       description,
       content,
       category,
       otherCategory,
+      subCategory,
       tags: tags.split(" "),
       imageUrl: fileUrl,
     };
 
-    // console.log(blogDetails)
     const update = await blogUpdate(blogId, BlogDetails, file ? false : true);
 
+    // If Validation Error
     if (update.error && update.formError) {
       return res.status(400).json({
         error: true,
         blogErrors: update.blogErrors,
+        success: false,
+      });
+    }
+
+    // If Error Occurred While Updating Blog
+    if (update?.serverError) {
+      return res.status(500).json({
+        serverError: true,
         success: false,
       });
     }
@@ -392,65 +529,104 @@ export const updateBlog = async (req, res) => {
   }
 };
 
+// Like or Unlike a Blog
 export const likeBlog = async (req, res) => {
   try {
     const { id } = req.params;
-    // console.log("Blog ID", id);
     const userId = req.user;
-    // console.log(userId);
+    let likesCount; 
+    
+    // Find Blog Based on the Id Provided
+    const validBlog = await findBlog(id);
+
+    // If Blog is Not Found
+    if (!validBlog.blog) {
+      return res.status(400).json({
+        error: true,
+        success: false,
+        message: "Blog Does Not Exists",
+      });
+    }
+
+    // If Error Occurred while Finding Blog
+    if (validBlog?.serverError) {
+      return res.status(400).json({
+        serverError: true,
+        success: false,
+      });
+    }
+
+    // If User Id is not Present
     if (!userId) {
       return res.status(401).json({
         error: true,
         auth: false,
-        message: "Please Login",
+        success: false,
       });
     }
-    const validBlog = await findBlog(id);
-    // console.log(validBlog);
-    if (!validBlog.blog) {
-      return res.status(400).json({
-        error: true,
-        message: "Blog Does Not Exists",
-      });
-    }
+
+    // Find User
     const validUser = await findUser(userId);
+
+    // If UserId does not exists
     if (!validUser) {
       return res.status(400).json({
         error: true,
+        success: false,
         auth: false,
-        message: "Please Login",
       });
     }
-    // console.log("Likes:", validBlog.blog.likes);
-    // console.log(validUser._id);
+
+    // If Error While Finding USer
+    if (validUser?.serverError) {
+      return res.status(500).json({
+        serverError: true,
+        success: false,
+      });
+    }
+
+    // Initial Like Count
+    likesCount = validBlog.blog.likes.length
+
+    // If UserId is not present in Likes Array, Add User ID
     if (
       validBlog.blog.likes.length === 0 ||
       validBlog.blog.likes.indexOf(validUser._id) === -1
     ) {
-      // console.log("Liking");
       validBlog.blog.likes = validBlog.blog.likes.push(validUser._id);
-      // console.log("new Array", validBlog.blog.likes);
-      await updateLikes(id, validBlog.blog.likes);
+
+      // Increment Likes Count
+      likesCount = likesCount + 1
+
+      // Update Likes
+      await updateLikes(id, validBlog.blog.likes, likesCount);
       return res.status(200).json({
         error: false,
         success: true,
-        message: "You Liked This Blog",
+        message: "Successfully Liked This Blog",
         liked: true,
+        likesCount
       });
     }
+
+    // If UserId is present in Likes Array, Remove it
     if (validBlog.blog.likes.includes(validUser._id)) {
-      // console.log("Disliking");
+
+      // Decrement likes Count
+      likesCount = likesCount - 1
+
       const filteredLikesArray = validBlog.blog.likes.filter(
         (like) => !like.equals(validUser._id)
-      );
-      console.log(filteredLikesArray);
-      // console.log(filteredLikesArray)
-      await updateLikes(id, filteredLikesArray);
+      )
+
+      // Update Likes
+      await updateLikes(id, filteredLikesArray, likesCount);
       return res.status(200).json({
         error: false,
         success: true,
         message: "You Disliked This Blog",
         liked: false,
+        likesCount
       });
     }
   } catch (error) {
